+++ assets/cards/card_abstraction.png
