+++ assets/cards/card_city_animated.gif
