+++ assets/cards/card_triangles.png
