+++ assets/cards/card_zebra.png
